face_detection_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
aiofiles==24.1.0
pydantic==2.9.2

opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
pillow==10.1.0
scrfd==0.2.1

torch==2.6.0+cu124
torchvision==0.21.0+cu124
--extra-index-url https://download.pytorch.org/whl/cu124
supervision==0.22.0

numpy==1.26.0

aiohttp==3.13.1
psutil==5.9.8